Prevent int underflow in IntegerQuantizer

+++ include/SZ3/quantizer/IntegerQuantizer.hpp
@@ -33,7 +33,8 @@ namespace SZ3 {
         // quantize the data with a prediction value, and returns the quantization index
         int quantize(T data, T pred) {
             T diff = data - pred;
-            int quant_index = (int) (fabs(diff) * this->error_bound_reciprocal) + 1;
+            // extend to long long before check against this->radius to prevent underflow
+            long long quant_index = (long long) (fabs(diff) * this->error_bound_reciprocal) + 1ll;
             if (quant_index < this->radius * 2) {
                 quant_index >>= 1;
                 int half_index = quant_index;
@@ -60,7 +61,8 @@ namespace SZ3 {
         // int quantize(T data, T pred, T& dec_data);
         int quantize_and_overwrite(T &data, T pred) {
             T diff = data - pred;
-            int quant_index = (int) (fabs(diff) * this->error_bound_reciprocal) + 1;
+            // extend to long long before check against this->radius to prevent underflow
+            long long quant_index = (long long) (fabs(diff) * this->error_bound_reciprocal) + 1ll;
             if (quant_index < this->radius * 2) {
                 quant_index >>= 1;
                 int half_index = quant_index;
@@ -88,7 +90,8 @@ namespace SZ3 {
 
         int quantize_and_overwrite(T ori, T pred, T &dest) {
             T diff = ori - pred;
-            int quant_index = (int) (fabs(diff) * this->error_bound_reciprocal) + 1;
+            // extend to long long before check against this->radius to prevent underflow
+            long long quant_index = (long long) (fabs(diff) * this->error_bound_reciprocal) + 1ll;
             if (quant_index < this->radius * 2) {
                 quant_index >>= 1;
                 int half_index = quant_index;
